# Fix GH-10648: add check function pointer into mbfl_encoding