--- a/ext/mbstring/libmbfl/filters/mbfilter_7bit.c
+++ b/ext/mbstring/libmbfl/filters/mbfilter_7bit.c
@@ -64,7 +64,8 @@ const mbfl_encoding mbfl_encoding_7bit = {
 	&vtbl_7bit_wchar,
 	&vtbl_wchar_7bit,
 	mb_7bit_to_wchar,
-	mb_wchar_to_7bit
+	mb_wchar_to_7bit,
+	NULL
 };
 
 #define CK(statement)	do { if ((statement) < 0) return (-1); } while (0)