--- a/ext/mbstring/libmbfl/filters/mbfilter_base64.c
+++ b/ext/mbstring/libmbfl/filters/mbfilter_base64.c
@@ -44,7 +44,8 @@ const mbfl_encoding mbfl_encoding_base64 = {
 	NULL,
 	NULL,
 	mb_base64_to_wchar,
-	mb_wchar_to_base64
+	mb_wchar_to_base64,
+	NULL
 };
 
 const struct mbfl_convert_vtbl vtbl_8bit_b64 = {