--- a/ext/mbstring/libmbfl/filters/mbfilter_qprint.c
+++ b/ext/mbstring/libmbfl/filters/mbfilter_qprint.c
@@ -46,7 +46,8 @@ const mbfl_encoding mbfl_encoding_qprint = {
 	NULL,
 	NULL,
 	mb_qprint_to_wchar,
-	mb_wchar_to_qprint
+	mb_wchar_to_qprint,
+	NULL
 };
 
 const struct mbfl_convert_vtbl vtbl_8bit_qprint = {