--- a/ext/mbstring/libmbfl/filters/mbfilter_ucs4.c
+++ b/ext/mbstring/libmbfl/filters/mbfilter_ucs4.c
@@ -56,7 +56,8 @@ const mbfl_encoding mbfl_encoding_ucs4 = {
 	&vtbl_ucs4_wchar,
 	&vtbl_wchar_ucs4,
 	mb_ucs4_to_wchar,
-	mb_wchar_to_ucs4be
+	mb_wchar_to_ucs4be,
+	NULL
 };
 
 const mbfl_encoding mbfl_encoding_ucs4be = {
@@ -69,7 +70,8 @@ const mbfl_encoding mbfl_encoding_ucs4be = {
 	&vtbl_ucs4be_wchar,
 	&vtbl_wchar_ucs4be,
 	mb_ucs4be_to_wchar,
-	mb_wchar_to_ucs4be
+	mb_wchar_to_ucs4be,
+	NULL
 };
 
 const mbfl_encoding mbfl_encoding_ucs4le = {
@@ -82,7 +84,8 @@ const mbfl_encoding mbfl_encoding_ucs4le = {
 	&vtbl_ucs4le_wchar,
 	&vtbl_wchar_ucs4le,
 	mb_ucs4le_to_wchar,
-	mb_wchar_to_ucs4le
+	mb_wchar_to_ucs4le,
+	NULL
 };
 
 const struct mbfl_convert_vtbl vtbl_ucs4_wchar = {