--- a/ext/mbstring/libmbfl/filters/mbfilter_utf16.c
+++ b/ext/mbstring/libmbfl/filters/mbfilter_utf16.c
@@ -49,7 +49,8 @@ const mbfl_encoding mbfl_encoding_utf16 = {
 	&vtbl_utf16_wchar,
 	&vtbl_wchar_utf16,
 	mb_utf16_to_wchar,
-	mb_wchar_to_utf16be
+	mb_wchar_to_utf16be,
+	NULL
 };
 
 const mbfl_encoding mbfl_encoding_utf16be = {
@@ -62,7 +63,8 @@ const mbfl_encoding mbfl_encoding_utf16be = {
 	&vtbl_utf16be_wchar,
 	&vtbl_wchar_utf16be,
 	mb_utf16be_to_wchar,
-	mb_wchar_to_utf16be
+	mb_wchar_to_utf16be,
+	NULL
 };
 
 const mbfl_encoding mbfl_encoding_utf16le = {
@@ -75,7 +77,8 @@ const mbfl_encoding mbfl_encoding_utf16le = {
 	&vtbl_utf16le_wchar,
 	&vtbl_wchar_utf16le,
 	mb_utf16le_to_wchar,
-	mb_wchar_to_utf16le
+	mb_wchar_to_utf16le,
+	NULL
 };
 
 const struct mbfl_convert_vtbl vtbl_utf16_wchar = {