--- a/ext/mbstring/libmbfl/filters/mbfilter_utf32.c
+++ b/ext/mbstring/libmbfl/filters/mbfilter_utf32.c
@@ -49,7 +49,8 @@ const mbfl_encoding mbfl_encoding_utf32 = {
 	&vtbl_utf32_wchar,
 	&vtbl_wchar_utf32,
 	mb_utf32_to_wchar,
-	mb_wchar_to_utf32be
+	mb_wchar_to_utf32be,
+	NULL
 };
 
 const mbfl_encoding mbfl_encoding_utf32be = {
@@ -62,7 +63,8 @@ const mbfl_encoding mbfl_encoding_utf32be = {
 	&vtbl_utf32be_wchar,
 	&vtbl_wchar_utf32be,
 	mb_utf32be_to_wchar,
-	mb_wchar_to_utf32be
+	mb_wchar_to_utf32be,
+	NULL
 };
 
 const mbfl_encoding mbfl_encoding_utf32le = {
@@ -75,7 +77,8 @@ const mbfl_encoding mbfl_encoding_utf32le = {
 	&vtbl_utf32le_wchar,
 	&vtbl_wchar_utf32le,
 	mb_utf32le_to_wchar,
-	mb_wchar_to_utf32le
+	mb_wchar_to_utf32le,
+	NULL
 };
 
 const struct mbfl_convert_vtbl vtbl_utf32_wchar = {