… ext/mbstring/libmbfl/filters/mbfilter_utf8_mobile.c
… ext/mbstring/libmbfl/filters/utf7_helper.h
… ext/mbstring/libmbfl/mbfl/mbfilter.c
… ext/mbstring/libmbfl/mbfl/mbfilter_8bit.c
… ext/mbstring/libmbfl/mbfl/mbfilter_pass.c
… ext/mbstring/libmbfl/mbfl/mbfilter_wchar.c
… ext/mbstring/libmbfl/mbfl/mbfl_encoding.h
… ext/mbstring/mbstring.c
… ext/mbstring/tests/gh10192_utf7.phpt
… ext/mbstring/tests/gh10192_utf7imap.phpt
… ext/mbstring/tests/gh10648.phpt
… ext/mbstring/tests/iso2022jp_encoding.phpt
… ext/mbstring/tests/utf_encodings.phpt