Previously, mbstring used the same logic for encoding validation as for encoding conversion.
However, there are cases where we want to use different logic for validation and conversion.
For example, if a string ends up with missing input required by the encoding, or if a character is input that is invalid as an encoding but can be converted, the conversion should succeed and the validation should fail.
To achieve this, a function pointer `mb_check_fn` has been added to `struct mbfl_encoding` to implement the logic used for validation.
Also, added implementation of validation logic for UTF-7, UTF7-IMAP, and JIS.