@alexdowad 2023-03-10
@pakutoma Great job for your first mbstring patch!!

I have left a few comments on where the patch can be improved. In general, we want to have *lots* of test cases which thoroughly explore the behavior of each function. I understand you may not want to spend a lot of time writing hundreds of test cases, but the number here is definitely a bit too small.

I would be even more comfortable if we used a fuzzer to check for any cases where there are unexpected changes in behavior, but again, I understand that learning to use a fuzzer right now might be a bigger step than what you want to take with your very first patch. I may try to make some time to fuzz this code when possible, and I don't think we need to hold up merging it for that.