@alexdowad 2023-03-11
@pakutoma I think you are making a lot of progress on this.

One other thing I remembered: I think this will change the behavior of `mb_detect_encoding` in non-strict detection mode. Or more to the point: I think it will *break* non-strict detection. (To enable non-strict detection, set the INI setting `mbstring.strict_detection` to false, or pass `false` as the 3rd argument to `mb_detect_encoding`.)

I think the best way to avoid that may be to guard the `for` loop which you added to `mb_detect_encoding` with `if (identd->strict)`.