@pakutoma 2023-03-11
@alexdowad Thanks for the lightning fast review!

By "break" I assume you mean that the added `for` loop increases the `num_illegalchars` for a given encoding, so that the `mbfl_encoding_detector_judge` function excludes the encoding.
Non-strict detection is expected to return more than one result, so in that sense this change is breaking it.
Hmmm, but in this sense, wasn't it originally broken when the input with `MBFL_BAD_INPUT` was passed? I think so, and that has been fixed in master.

And looking at master, I see that instead of increasing `num_illegalchars`, it adds `1000` to `demerits`.
https://github.com/php/php-src/blob/master/ext/mbstring/mbstring.c#L3053-3060

I think I can add a number like `1000` to `data->score` instead of guarding the `for` loop with `if`, but is that a bit hacky?
I think the advantage is that we can use the output of the `check` function as a heuristic for non-strict detection.