@alexdowad 2023-03-11
@pakutoma About non-strict detection... first, let me say that this a dubious feature, which probably should not have been included. But since it was, and we already have people using it, we need to make it continue working.

Here is what non-strict detection does: normally, if *all* the candidate text encodings are invalid, then `mb_detect_encoding` does not return any encoding. *However*, if non-strict detection is enabled, then as soon as N-1 candidate encodings are eliminated, then `mb_detect_encoding` immediately returns the last remaining one.

An example: Let's say we have a 1000-byte string as input. Maybe the three candidate encodings are UTF-8, ISO-2022-JP, and EUC-JP. Imagine that the 100th byte is invalid in UTF-8, the 200th byte is invalid in ISO-2022-JP, and the 300th byte is invalid in EUC-JP. In that case, with *non-strict* detection, the return value will be `"EUC-JP"`. But with *strict* detection, the return value will be null.

You will probably agree that this is a strange feature, and that it is probably a bad idea for people to use it.

Do you see why your new check functions are a problem for that feature? With the new check functions, we have no way to determine that *this* encoding is invalid on the 100th byte, but *that* encoding is invalid on the 200th byte. The check functions just process the entire string and return true or false.

This is actually the reason why I originally said that the function signature should include a state argument. But I think we can get around the problem just using a guard `if` clause, as I said above. This means that non-strict detection will not benefit from your check functions; but anyways, non-strict detection is... well, *non-strict*. So I don't think it matters if non-strict detection does not pay attention to ISO-2022-JP strings ending in an incorrect mode.