@pakutoma 2023-03-11
@alexdowad 
Thanks for the detailed explanation, it helped me to understand.
It seems that the check function needs a major change to "fully" support this feature and there is not much to be gained by it.

However, considering that adding `data->score += 1000` gives a slightly better result for non-strict detection than doing nothing, I think it's a good idea to add this, but I also understand that you don't want to add more code for non-strict detection. What do you think?