@alexdowad 2023-03-13
@pakutoma Sorry that my response wasn't "lightning fast" this time!

At the moment, my personal preference is that we leave non-strict detection alone and just allow your new check functions to help improve the accuracy of strict detection. I am always open to being convinced otherwise if there are good reasons.

@pakutoma One more thing. Could you please add entries in NEWS and UPGRADING for this change?

I am going on a trip tomorrow, but when I get to my hotel and have some extra time, I may try to build your code and run the tests.