@pakutoma 2023-03-14
@alexdowad Thanks! I'm working slowly so don't mind me. Have a great trip!

> my personal preference is that we leave non-strict detection alone and just allow your new check functions to help improve the accuracy of strict detection.
Thanks, I will do this.
Surely it would be better to allow users to select character codes in other appropriate ways rather than to continue to have them use unstable methods by making ad hoc fixes to non-strict detection methods.

> Could you please add entries in NEWS and UPGRADING for this change?
I will try to write them.
Is the change to write in UPGRADING only due to this PR?
As far as I know, mb_(check|detect|convert)_encoding has also changed significantly in PHP 8.1,  do I able to include that change as well?