@alexdowad 2023-03-14
> > Could you please add entries in NEWS and UPGRADING for this change?
> 
> I will try to write them. Is the change to write in UPGRADING only due to this PR? As far as I know, mb_(check|detect|convert)_encoding has also changed significantly in PHP 8.1, do I able to include that change as well?

When I started working on mbstring, I should have included entries in UPGRADING at that time, but wasn't aware of the use of that file yet.

Whatever is entered in UPGRADING now will go into the upgrade notes for the next release, so I think it should only include changes which are relevant to people upgrading from the previous release to this one.