@alexdowad 2023-03-14
Just built the code; indeed, all the tests pass. That is good to see.

Let me request a few more tweaks:
- Squash into a single commit.
- Remove trailing spaces in NEWS.
- UPGRADING mentions mb_convert_encoding(), but I think that might be incorrect; does this patch affect mb_convert_encoding??
- For UTF-7, don't say `if (cp >= ...) return true; else return false;`. Just say `return (cp >= ...);`

I still need to:
- Carefully examine the new test cases and check if the EXPECT section looks right.
- Benchmark to make sure there is no significant performance loss.
- Fuzz.

@youkidearitai pointed out in another thread that this adjustment to `mb_check_encoding` should probably apply to ISO-2022-JP-2004 as well. You could include that in this PR or make a separate one after this one is merged; either way is fine.

@cmb69 @Girgias Do you have some comments? Do you think it is appropriate to target 8.2 with this PR?