@alexdowad 2023-03-18
Little script I made to benchmark `mb_check_encoding`: https://gist.github.com/alexdowad/b7952a8d733ba9255daa527d15004928

Results:

```
new|mb_check_encoding|JIS, short (0-5 chars)|0.00029759407043457|10000
new|mb_check_encoding|JIS, medium (~100 chars)|0.0019969940185547|10000
new|mb_check_encoding|JIS, long (~10000 chars)|0.1247300306956|10000
old|mb_check_encoding|JIS, short (0-5 chars)|0.00040802955627441|10000
old|mb_check_encoding|JIS, medium (~100 chars)|0.0023880004882812|10000
old|mb_check_encoding|JIS, long (~10000 chars)|0.21371102333069|10000
```

This is a very nice speedup for `mb_check_encoding`.